case_studies/supreme-court-ai-hackathon-2024/scij-skillsets/Python/EnrichmentSkills/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
azure-identity

urllib3